Various improvements
*Improved CMakeLists.txt to include correct compiler flags
*FIX warnings in LegendrePolynomial and LegendrePolynomialsExplicit
*New versions of submodules CTIOL and Catch2
*Improved testing
*Added test SquareMatrix.test.cpp that tests LU decomposition

+++ include/FUN/LegendrePolynomial.h
@@ -85,7 +85,7 @@ namespace util {
                 return 1;
 
             int result = n;
-            for(int i = 2; i <= k; ++i)
+            for(size_t i = 2; i <= k; ++i)
             {
                 result *= (n - i + 1);
                 result /= i;
